Properly authroize livewire RBAC controller,s

+++ src/Admin/Http/Livewire/RBAC/DeleteRole.php
@@ -2,14 +2,20 @@
 
 namespace SquadMS\Foundation\Admin\Http\Livewire\RBAC;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Spatie\Permission\Models\Role;
 use SquadMS\Foundation\Admin\Http\Livewire\Contracts\AbstractModalComponent;
 
 class DeleteRole extends AbstractModalComponent
 {
+    use AuthorizesRequests;
+
     public Role $role;
 
     public function deleteRole() {
+        /* Authorize the action */
+        $this->authorize('delete', $this->role);
+
         /* Delete the Role */
         $this->role->delete();
